RSSEC-62 refactory: move part of the code to another function
